fix double->int64_t warnings in LAVPMovie.m

+++ libavPlayer/LAVPMovie.h
@@ -32,7 +32,7 @@
 @interface LAVPMovie : NSObject {
 @private
     VideoState *is;
-    double lastPosition;
+    int64_t lastPosition;
 }
 
 @property (retain, readonly) NSURL *url;
